Definindo services para Aeronave e Modelo

+++ ABCDAirlinesAPI/ABCDAir/Models/Aeronave.cs
@@ -10,8 +10,8 @@ namespace ABCDAir.Models
     {
         [Key]
         public string PREFIX { get; set; }
-        public double MAX_DEPARTURE_WEIGHT { get; set; }
-        public double MAX_LANDING_WEIGHT { get; set; }
+        public decimal MAX_DEPARTURE_WEIGHT { get; set; }
+        public decimal MAX_LANDING_WEIGHT { get; set; }
         public bool ACTIVE { get; set; }
         public string AIRCRAFT_MODEL { get; set; }
     }
